envp added to minishell as a linked list. should be able to modify environment variables now

+++ libft/ft_lstsize.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstsize_bonus.c                                 :+:      :+:    :+:   */
+/*   ft_lstsize.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: cwon <cwon@student.42bangkok.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/06 08:35:47 by cwon              #+#    #+#             */
-/*   Updated: 2024/09/08 17:37:52 by cwon             ###   ########.fr       */
+/*   Updated: 2025/05/25 19:48:34 by cwon             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (lst)
